+++ src/test/php/net/xp_framework/unittest/core/generics/GenericTypesTest.class.php
@@ -1,24 +1,23 @@
 <?php namespace net\xp_framework\unittest\core\generics;
 
-use lang\Primitive;
-use lang\GenericTypes;
-use lang\XPClass;
+use lang\{GenericTypes, Primitive, XPClass};
+use unittest\TestCase;
 
 /**
  * TestCase for lang.GenericTypes
  */
-class GenericTypesTest extends \unittest\TestCase {
+class GenericTypesTest extends TestCase {
   private static $filter;
 
   #[@beforeClass]
   public static function defineBase() {
-    self::$filter= XPClass::forName('net.xp_framework.unittest.core.generics.NSFilter');
+    self::$filter= XPClass::forName('net.xp_framework.unittest.core.generics.ArrayFilter');
   }
   
   #[@test]
   public function newType0_returns_literal() {
     $this->assertEquals(
-      'net\xp_framework\unittest\core\generics\NSFilter··þint',
+      "net\\xp_framework\\unittest\\core\\generics\\ArrayFilter\xb7\xb7\xfeint",
       (new GenericTypes())->newType0(self::$filter, [Primitive::$INT])
     );
   }